refactor(button): simplify styles and add onClick handling
Refactored the button style definitions to simplify the default class name logic. Added support for `onClick` event handling with proper disabled state prevention. Updated story files and adjusted tests for consistency with the new structure.
